added Abstractmethod

+++ User/UserManager.py
@@ -20,11 +20,11 @@ class UserManager:
     def get_email(self, user_id):
         return self.__user_storage.get_email(user_id)
 
-    # def add_new_user(self, user):
-    #     self.__user_storage.put_one(user)
-
     def delete_user(self, user_id):
-        self.__user_storage.delete_one(user_id)
+        if user_id > len(self.get_all_users):
+            print('Пользователь с таким user_id не найден')
+        else:
+            self.__user_storage.delete_one(user_id)
 
     def add_new_user(self, user: User):
         a = []
@@ -32,8 +32,6 @@ class UserManager:
         for i in range(len(self.get_all_users)):
             a.append(self.get_login(i))
             b.append(self.get_email(i))
-        # print(a)
-        # print(b)
         if user.login in a:
             print('Пользователь с таким login уже существует')
         elif user.email in b:
